# ENH: Release the GIL while computing KDE kernel estimate