<!-- 
Thanks for contributing a pull request! Please ensure that
your PR satisfies the checklist before submitting:
http://scipy.github.io/devdocs/dev/contributor/development_workflow.html#checklist-before-submitting-a-pr

Also, please name and describe your PR as you would write a
commit message:
http://scipy.github.io/devdocs/dev/contributor/development_workflow.html#writing-the-commit-message

Note that we are a team of volunteers; we appreciate your
patience during the review process.

Again, thanks for contributing!
-->

#### What does this implement/fix?
The computationally intensive part of the `scipy.stats.gaussian_kde.evaluate` involves calling `gaussian_kernel_estimate`, which has a triply nested for-loop, but does not release the GIL. This loop could be parallelized by OpenMP as well, but lacking a guide on when and how `scipy` controls this, I opted to just release the GIL and leave it to the user to call it from multiple threads.

This should have a negligible impact on existing code but enable future code to evaluate KDEs while other code occupies the GIL. I doubled the speed of my algorithm using three threads using `concurrent.futures.ThreadPoolExecutor` instead of the singly threaded implementation. 

#### Additional information
I'm unable to build `scipy` from source locally, and hope that the continuous integration system can exercise the full test suite. I have extracted the logic into a separate Cython module and tested it in isolation.

This change does open the possibility that user code might mutate the memory buffers behind the attributes of the `scipy.stats.gaussian_kde` object while concurrently using those buffers in another thread. Fortunately, no API code appears to modify attributes in-place.