@mobiusklein 2022-08-02
@tupui I moved the loops to a `cdef` function that runs without the GIL and just invoke that inside the nogil block. This may be a little bit slower in the trivial case because now we have to pass and acquire five buffers, but shouldn't make the performance that much worse.