@jjerphan 2022-08-02
To answer your questions, @tupui:

> I am wondering if we could not put more operations in the nogil statement.

Calling the numpy Python API and raising errors:
https://github.com/scipy/scipy/blob/88e68004b078c6d503e5c77a09d144decedee7b7/scipy/stats/_stats.pyx#L742-L751

necessitate interacting with the CPython interpreter. Hence no other operation in `gaussian_kernel_estimate` can be move in the `nogil` context.

> And also maybe moving the hot part to a cdef could bring additional performance?

That's right. It's a necessary condition for optimal performance but not a sufficient one (in particular it's still possible to be bound to the GIL in a `cdef` function).