@tupui 2022-08-02
Thanks @jjerphan for the precisions.

About moving more things, I was referring to the dot products and the type casting. But yes if that's better to keep this in Python no problem.

```python
points_ = np.dot(points, whitening).astype(dtype, copy=False) 
xi_ = np.dot(xi, whitening).astype(dtype, copy=False) 
values_ = values.astype(dtype, copy=False)
```