@jjerphan 2022-08-02
Ah yes, it's possible to perform them in a `nogil` context (e.g. via calls to BLAS directly) but one still need to allocate buffers for them, so keeping those in Python makes sense.