@mobiusklein 2022-08-30
@mdhaber 
The `cho_cov_ = cho_cov.astype(dtype, copy=False)` step is correct. The conversion logic in this function is obtuse because Cython can't infer fused type specializations from buffer types and we are passing five buffers which can all be different numerical types with different widths but want to do all of the computation on a single type.

I can't actually _build_ `scipy` locally, so unless https://dev.azure.com/scipy-org/SciPy/_build/results?buildId=20229&view=logs&j=ede89041-5e8e-5201-ce5c-f68c2cbf02dd&t=88295f8e-e8a1-53ce-523a-c740e2f72a8d runs the `asv` tests you are referring to, I can't check this formally.

If that is a blocker, what are alternatives?