@mdhaber 2022-08-30
> If that is a blocker, what are alternatives?

Well, we need to see something that demonstrates that this is an improvement. [Here](https://github.com/scipy/scipy/blob/cc63003326b7ae2e9dd41cd6af0e6e3079c32b95/benchmarks/benchmarks/stats.py#L431) is the benchmark code. As an executable snippet, it's something like this:

```python3
import numpy as np
from scipy import stats

rng = np.random.default_rng(12345678)
n = 2000
m1 = rng.normal(size=n)
m2 = rng.normal(scale=0.5, size=n)
xmin = m1.min()
xmax = m1.max()
ymin = m2.min()
ymax = m2.max()

X, Y = np.mgrid[xmin:xmax:80j, ymin:ymax:80j]
positions = np.vstack([X.ravel(), Y.ravel()])
values = np.vstack([m1, m2])
kernel = stats.gaussian_kde(values)

kernel.pdf(positions)
```

Timing just that last line using `%timeit`, I get:
```
137 ms ± 147 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  # main
147 ms ± 211 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  # PR
```

The benchmark size was reduced recently, and I thought it might need a larger size to shine, so I bumped it back up to the original:
```python3
X, Y = np.mgrid[xmin:xmax:200j, ymin:ymax:200j]
```

And I see:
```
852 ms ± 855 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)  # main
922 ms ± 1.57 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # PR
```

Is there a larger size problem at which I'd expect to see a performance improvement, or will I just not benefit from this on my machine?
I'm curious what others get here. 

|      | 80x80  | 200x200 |
|------|--------|---------|
| main | 137 ms    | 852 ms    |
| PR   | 147 ms | 922 ms  |