@mobiusklein 2022-09-02
On its own, my changes don't make `gaussian_kde.evaluate` faster. What they do is release the global interpreter lock while doing the time consuming triply nested for-loop of numerical operations so that while that's happening, another thread can occupy the Python interpreter (or that function may be called from multiple threads).

```python
import numpy as np
from scipy import stats

rng = np.random.default_rng(12345678)
n = 2000
m1 = rng.normal(size=n)
m2 = rng.normal(scale=0.5, size=n)
xmin = m1.min()
xmax = m1.max()
ymin = m2.min()
ymax = m2.max()

X, Y = np.mgrid[xmin:xmax:80j, ymin:ymax:80j]
positions = np.vstack([X.ravel(), Y.ravel()])

values = np.vstack([m1, m2])
kernel = stats.gaussian_kde(values)

from isolated_c_extension import evaluate_gaussian_kde

def reference():
    return kernel.pdf(positions)

def single_thread_modified():
    return evaluate_gaussian_kde(kernel, positions)

%timeit reference()
%timeit single_thread_modified()
```
These singly threaded executions on my box give me:
```
# master, singly threaded
131 ms ± 922 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
# PR, singly threaded
141 ms ± 654 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
Now, I spin up a small thread pool in Python
```python
from concurrent import futures

executor = futures.ThreadPoolExecutor(3)

def multi_thread_reference():
    jobs = []
    step_size = positions.shape[1] // 3
    for i in range(3):
        chunk = positions[:, i * step_size:(i + 1) * step_size]
        jobs.append(
            executor.submit(
                kernel.pdf,
                chunk
            )
        )
    return [j.result() for j in jobs]

def multi_thread_modified():
    jobs = []
    step_size = positions.shape[1] // 3
    for i in range(3):
        chunk = positions[:, i * step_size:(i + 1) * step_size]
        jobs.append(
            executor.submit(
                evaluate_gaussian_kde,
                kernel,
                chunk
            )
        )
    return [j.result() for j in jobs]

%timeit multi_thread_reference()

%timeit multi_thread_modified()
```
Again, timing on my box with that thread pool (which drains completely between each execution of the benchmark)
```
# master, three threads
130 ms ± 151 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
# PR, three threads
48.5 ms ± 631 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
So by running the function in chunks over three threads I get a ~260% speed up on the same data. This never quite reaches the # of threads because the memory allocation and buffer type conversion operations that are done up-front are still holding the GIL.

In my use case, I actually have four independent KDEs I need to evaluate dozens of times per iteration on a potentially large (10k+ points) dataset, this lets me evaluate them all concurrently rather than wait for each to run in sequence.