@mdhaber 2022-09-02
OK, I see now. To me that sounds like enough of a potential improvement - to those who seek it - to justify the slight regression in the usual case. @tupui @jjerphan please let me know if you think that's right. In the meantime, I'll go ahead and merge.