@jjerphan 2022-09-02
LGTM. :+1: 

Potentially, there might be subscription with the above proposal proposed in https://github.com/scipy/scipy/pull/16758#issuecomment-1234985423 if it is being used in user-code with another layer of parallelism.

[`threadpoolctl`](https://github.com/joblib/threadpoolctl/) might help prevent such problems, but this would add another dependency to SciPy.