@tupui 2022-09-02
> [`threadpoolctl`](https://github.com/joblib/threadpoolctl/) might help prevent such problems, but this would add another dependency to SciPy.

It might yes. We are discussing parallelism architecture in general as part of the NASA Roses Grant (Thomas from scikit-learn is there as well). So let's wait for this group to make a community RFC.